tavilytest.py
from tavily import TavilyClient
import os
from dotenv import load_dotenv

load_dotenv()
client = TavilyClient(api_key=os.getenv("TAVILY_API_KEY"))
resp = client.search("Indian fintech market size 2026")
print(resp)
