requirements.txt
langgraph==0.0.45
langchain-openai==0.1.7
langchain-community==0.2.6
tavily-python==0.3.3
python-telegram-bot==21.5
python-dotenv==1.0.1
openai==1.12.0

flask

